+++ resources/views/filament/standard/components/brand.blade.php
@@ -1,0 +1,1 @@
+<x-public.brand class="dc-brand" />