+++ resources/views/filament/standard/components/support.blade.php
@@ -1,0 +1,5 @@
+<aside class="dc-support">
+    <x-heroicon-o-light-bulb />
+    <div><strong>{{ __('dashboard.support_title') }}</strong><p>{{ __('dashboard.support_hint') }}</p></div>
+    <a href="{{ config('app.footer.issues') }}" target="_blank" rel="noopener noreferrer">{{ __('dashboard.support_link') }} <x-heroicon-o-arrow-right /></a>
+</aside>